pyproject.toml
[project]
name = "welding-maintenance-demo"
version = "0.2.0"
description = "Vision + sensor predictive maintenance demo for a robotic welding arm (PyTorch)"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "numpy>=1.26.0",
  "pandas>=2.3.0",
  "scikit-learn>=1.4.0",
  "matplotlib>=3.8.0",
  "seaborn>=0.13.0",
  "scipy>=1.11.0",
  "torch>=2.3.0",
  "torchvision>=0.18.0",
  "jupyter>=1.0.0",
  "joblib>=1.3.0",
  "Pillow>=10.0.0"
]

[project.optional-dependencies]
dev = [
  "black>=23.0.0",
  "flake8>=6.0.0",
  "pytest>=7.0.0"
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.flake8]
max-line-length = 100
extend-ignore = "E203"
